+++ js/util.js
@@ -1,0 +1,20 @@
+const generateId = () => {
+  let currentId = 0;
+  return function() {
+    currentId++;
+    return currentId;
+  };
+};
+
+function getRandomPositiveInteger(a, b) {
+  if (a < 0 || b < 0) {
+    return NaN;
+  }
+  const lower = Math.ceil(Math.min(a, b));
+  const upper = Math.floor(Math.max(a, b));
+  const result = Math.round(Math.random() * (upper - lower) + lower);
+  return Math.floor(result);
+}
+const getRandomArrayElement = (elements) => elements[getRandomPositiveInteger(0, elements.length - 1)];
+
+export {generateId, getRandomPositiveInteger, getRandomArrayElement};
